Memperbaiki tabel nomor wa

+++ database/migrations/2024_05_15_132545_create_bookings_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId("car_id");
             $table->string("nama_lengkap");
             $table->string("alamat_lengkap");
-            $table->integer("nomor_handphone");
+            $table->bigInteger("nomor_handphone");
             $table->timestamps();
         });
     }
